delete api fix

+++ src/app/api/book/deletebook/route.ts
@@ -7,7 +7,24 @@ export async function POST(request: NextRequest) {
   if (!id) {
     return NextResponse.json({ error: "id is required" }, { status: 400 });
   }
+
   try {
+    const exchangeRequests = await prisma.exchangeRequest.findMany({
+      where: {
+        OR: [{ requestedBookId: id }, { offeredBookId: id }],
+      },
+    });
+
+    if (exchangeRequests.length > 0) {
+      return NextResponse.json(
+        {
+          error:
+            "Book cannot be deleted because it is involved in active exchange requests.",
+        },
+        { status: 400 }
+      );
+    }
+
     const book = await prisma.book.delete({
       where: {
         id: id,
